Refactor(Domain) : RevokeApplicationUseCase.kt   - return type 을 RevokeResult 로 수정   - result 에 따라 어떻게 처리할 것인지 판단할 수 있도록 수정   - exception 별로 처리 추가해야함

+++ domain/src/main/java/com/gyleedev/githubsearch/domain/usecase/RevokeApplicationUseCase.kt
@@ -1,13 +1,20 @@
 package com.gyleedev.githubsearch.domain.usecase
 
+import com.gyleedev.githubsearch.domain.model.RevokeResult
 import com.gyleedev.githubsearch.domain.repository.GitHubRepository
 import javax.inject.Inject
 
 class RevokeApplicationUseCase @Inject constructor(
     private val repository: GitHubRepository,
 ) {
-    suspend operator fun invoke() {
-        repository.revokeApplication()
-        repository.deleteAccessToken()
+    suspend operator fun invoke(): RevokeResult = try {
+        val result = repository.revokeApplication()
+        if (result == RevokeResult.SUCCESS) {
+            repository.deleteAccessToken()
+        }
+        result
+    } catch (e: Exception) {
+        // 어떤 실패인지 추가할 필요 있음
+        RevokeResult.FAIL
     }
 }
